remove generated command config classes
