+++ packages/patrol_cli/lib/src/base/extensions/core.dart
