--- packages/patrol_cli/lib/src/commands/develop_command.freezed.dart
@@ -1,391 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'develop_command.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-mixin _$DevelopCommandConfig {
-  Device get device => throw _privateConstructorUsedError;
-  String get target => throw _privateConstructorUsedError;
-  Map<String, String> get dartDefines => throw _privateConstructorUsedError;
-  bool get displayLabel => throw _privateConstructorUsedError;
-  bool get uninstall =>
-      throw _privateConstructorUsedError; // Android-only options
-  String? get packageName => throw _privateConstructorUsedError;
-  String? get androidFlavor =>
-      throw _privateConstructorUsedError; // iOS-only options
-  String? get bundleId => throw _privateConstructorUsedError;
-  String? get iosFlavor => throw _privateConstructorUsedError;
-  String get scheme => throw _privateConstructorUsedError;
-  String get xcconfigFile => throw _privateConstructorUsedError;
-  String get configuration => throw _privateConstructorUsedError;
-
-  @JsonKey(ignore: true)
-  $DevelopCommandConfigCopyWith<DevelopCommandConfig> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $DevelopCommandConfigCopyWith<$Res> {
-  factory $DevelopCommandConfigCopyWith(DevelopCommandConfig value,
-          $Res Function(DevelopCommandConfig) then) =
-      _$DevelopCommandConfigCopyWithImpl<$Res, DevelopCommandConfig>;
-  @useResult
-  $Res call(
-      {Device device,
-      String target,
-      Map<String, String> dartDefines,
-      bool displayLabel,
-      bool uninstall,
-      String? packageName,
-      String? androidFlavor,
-      String? bundleId,
-      String? iosFlavor,
-      String scheme,
-      String xcconfigFile,
-      String configuration});
-}
-
-/// @nodoc
-class _$DevelopCommandConfigCopyWithImpl<$Res,
-        $Val extends DevelopCommandConfig>
-    implements $DevelopCommandConfigCopyWith<$Res> {
-  _$DevelopCommandConfigCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? device = null,
-    Object? target = null,
-    Object? dartDefines = null,
-    Object? displayLabel = null,
-    Object? uninstall = null,
-    Object? packageName = freezed,
-    Object? androidFlavor = freezed,
-    Object? bundleId = freezed,
-    Object? iosFlavor = freezed,
-    Object? scheme = null,
-    Object? xcconfigFile = null,
-    Object? configuration = null,
-  }) {
-    return _then(_value.copyWith(
-      device: null == device
-          ? _value.device
-          : device // ignore: cast_nullable_to_non_nullable
-              as Device,
-      target: null == target
-          ? _value.target
-          : target // ignore: cast_nullable_to_non_nullable
-              as String,
-      dartDefines: null == dartDefines
-          ? _value.dartDefines
-          : dartDefines // ignore: cast_nullable_to_non_nullable
-              as Map<String, String>,
-      displayLabel: null == displayLabel
-          ? _value.displayLabel
-          : displayLabel // ignore: cast_nullable_to_non_nullable
-              as bool,
-      uninstall: null == uninstall
-          ? _value.uninstall
-          : uninstall // ignore: cast_nullable_to_non_nullable
-              as bool,
-      packageName: freezed == packageName
-          ? _value.packageName
-          : packageName // ignore: cast_nullable_to_non_nullable
-              as String?,
-      androidFlavor: freezed == androidFlavor
-          ? _value.androidFlavor
-          : androidFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      bundleId: freezed == bundleId
-          ? _value.bundleId
-          : bundleId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      iosFlavor: freezed == iosFlavor
-          ? _value.iosFlavor
-          : iosFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      scheme: null == scheme
-          ? _value.scheme
-          : scheme // ignore: cast_nullable_to_non_nullable
-              as String,
-      xcconfigFile: null == xcconfigFile
-          ? _value.xcconfigFile
-          : xcconfigFile // ignore: cast_nullable_to_non_nullable
-              as String,
-      configuration: null == configuration
-          ? _value.configuration
-          : configuration // ignore: cast_nullable_to_non_nullable
-              as String,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$_DevelopCommandConfigCopyWith<$Res>
-    implements $DevelopCommandConfigCopyWith<$Res> {
-  factory _$$_DevelopCommandConfigCopyWith(_$_DevelopCommandConfig value,
-          $Res Function(_$_DevelopCommandConfig) then) =
-      __$$_DevelopCommandConfigCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call(
-      {Device device,
-      String target,
-      Map<String, String> dartDefines,
-      bool displayLabel,
-      bool uninstall,
-      String? packageName,
-      String? androidFlavor,
-      String? bundleId,
-      String? iosFlavor,
-      String scheme,
-      String xcconfigFile,
-      String configuration});
-}
-
-/// @nodoc
-class __$$_DevelopCommandConfigCopyWithImpl<$Res>
-    extends _$DevelopCommandConfigCopyWithImpl<$Res, _$_DevelopCommandConfig>
-    implements _$$_DevelopCommandConfigCopyWith<$Res> {
-  __$$_DevelopCommandConfigCopyWithImpl(_$_DevelopCommandConfig _value,
-      $Res Function(_$_DevelopCommandConfig) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? device = null,
-    Object? target = null,
-    Object? dartDefines = null,
-    Object? displayLabel = null,
-    Object? uninstall = null,
-    Object? packageName = freezed,
-    Object? androidFlavor = freezed,
-    Object? bundleId = freezed,
-    Object? iosFlavor = freezed,
-    Object? scheme = null,
-    Object? xcconfigFile = null,
-    Object? configuration = null,
-  }) {
-    return _then(_$_DevelopCommandConfig(
-      device: null == device
-          ? _value.device
-          : device // ignore: cast_nullable_to_non_nullable
-              as Device,
-      target: null == target
-          ? _value.target
-          : target // ignore: cast_nullable_to_non_nullable
-              as String,
-      dartDefines: null == dartDefines
-          ? _value._dartDefines
-          : dartDefines // ignore: cast_nullable_to_non_nullable
-              as Map<String, String>,
-      displayLabel: null == displayLabel
-          ? _value.displayLabel
-          : displayLabel // ignore: cast_nullable_to_non_nullable
-              as bool,
-      uninstall: null == uninstall
-          ? _value.uninstall
-          : uninstall // ignore: cast_nullable_to_non_nullable
-              as bool,
-      packageName: freezed == packageName
-          ? _value.packageName
-          : packageName // ignore: cast_nullable_to_non_nullable
-              as String?,
-      androidFlavor: freezed == androidFlavor
-          ? _value.androidFlavor
-          : androidFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      bundleId: freezed == bundleId
-          ? _value.bundleId
-          : bundleId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      iosFlavor: freezed == iosFlavor
-          ? _value.iosFlavor
-          : iosFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      scheme: null == scheme
-          ? _value.scheme
-          : scheme // ignore: cast_nullable_to_non_nullable
-              as String,
-      xcconfigFile: null == xcconfigFile
-          ? _value.xcconfigFile
-          : xcconfigFile // ignore: cast_nullable_to_non_nullable
-              as String,
-      configuration: null == configuration
-          ? _value.configuration
-          : configuration // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$_DevelopCommandConfig implements _DevelopCommandConfig {
-  const _$_DevelopCommandConfig(
-      {required this.device,
-      required this.target,
-      required final Map<String, String> dartDefines,
-      required this.displayLabel,
-      required this.uninstall,
-      required this.packageName,
-      required this.androidFlavor,
-      required this.bundleId,
-      required this.iosFlavor,
-      required this.scheme,
-      required this.xcconfigFile,
-      required this.configuration})
-      : _dartDefines = dartDefines;
-
-  @override
-  final Device device;
-  @override
-  final String target;
-  final Map<String, String> _dartDefines;
-  @override
-  Map<String, String> get dartDefines {
-    if (_dartDefines is EqualUnmodifiableMapView) return _dartDefines;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableMapView(_dartDefines);
-  }
-
-  @override
-  final bool displayLabel;
-  @override
-  final bool uninstall;
-// Android-only options
-  @override
-  final String? packageName;
-  @override
-  final String? androidFlavor;
-// iOS-only options
-  @override
-  final String? bundleId;
-  @override
-  final String? iosFlavor;
-  @override
-  final String scheme;
-  @override
-  final String xcconfigFile;
-  @override
-  final String configuration;
-
-  @override
-  String toString() {
-    return 'DevelopCommandConfig(device: $device, target: $target, dartDefines: $dartDefines, displayLabel: $displayLabel, uninstall: $uninstall, packageName: $packageName, androidFlavor: $androidFlavor, bundleId: $bundleId, iosFlavor: $iosFlavor, scheme: $scheme, xcconfigFile: $xcconfigFile, configuration: $configuration)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$_DevelopCommandConfig &&
-            (identical(other.device, device) || other.device == device) &&
-            (identical(other.target, target) || other.target == target) &&
-            const DeepCollectionEquality()
-                .equals(other._dartDefines, _dartDefines) &&
-            (identical(other.displayLabel, displayLabel) ||
-                other.displayLabel == displayLabel) &&
-            (identical(other.uninstall, uninstall) ||
-                other.uninstall == uninstall) &&
-            (identical(other.packageName, packageName) ||
-                other.packageName == packageName) &&
-            (identical(other.androidFlavor, androidFlavor) ||
-                other.androidFlavor == androidFlavor) &&
-            (identical(other.bundleId, bundleId) ||
-                other.bundleId == bundleId) &&
-            (identical(other.iosFlavor, iosFlavor) ||
-                other.iosFlavor == iosFlavor) &&
-            (identical(other.scheme, scheme) || other.scheme == scheme) &&
-            (identical(other.xcconfigFile, xcconfigFile) ||
-                other.xcconfigFile == xcconfigFile) &&
-            (identical(other.configuration, configuration) ||
-                other.configuration == configuration));
-  }
-
-  @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      device,
-      target,
-      const DeepCollectionEquality().hash(_dartDefines),
-      displayLabel,
-      uninstall,
-      packageName,
-      androidFlavor,
-      bundleId,
-      iosFlavor,
-      scheme,
-      xcconfigFile,
-      configuration);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$_DevelopCommandConfigCopyWith<_$_DevelopCommandConfig> get copyWith =>
-      __$$_DevelopCommandConfigCopyWithImpl<_$_DevelopCommandConfig>(
-          this, _$identity);
-}
-
-abstract class _DevelopCommandConfig implements DevelopCommandConfig {
-  const factory _DevelopCommandConfig(
-      {required final Device device,
-      required final String target,
-      required final Map<String, String> dartDefines,
-      required final bool displayLabel,
-      required final bool uninstall,
-      required final String? packageName,
-      required final String? androidFlavor,
-      required final String? bundleId,
-      required final String? iosFlavor,
-      required final String scheme,
-      required final String xcconfigFile,
-      required final String configuration}) = _$_DevelopCommandConfig;
-
-  @override
-  Device get device;
-  @override
-  String get target;
-  @override
-  Map<String, String> get dartDefines;
-  @override
-  bool get displayLabel;
-  @override
-  bool get uninstall;
-  @override // Android-only options
-  String? get packageName;
-  @override
-  String? get androidFlavor;
-  @override // iOS-only options
-  String? get bundleId;
-  @override
-  String? get iosFlavor;
-  @override
-  String get scheme;
-  @override
-  String get xcconfigFile;
-  @override
-  String get configuration;
-  @override
-  @JsonKey(ignore: true)
-  _$$_DevelopCommandConfigCopyWith<_$_DevelopCommandConfig> get copyWith =>
-      throw _privateConstructorUsedError;
-}
