--- packages/patrol_cli/lib/src/commands/test_command.freezed.dart
@@ -1,425 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'test_command.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-mixin _$TestCommandConfig {
-  List<Device> get devices => throw _privateConstructorUsedError;
-  List<String> get targets => throw _privateConstructorUsedError;
-  Map<String, String> get dartDefines => throw _privateConstructorUsedError;
-  int get repeat => throw _privateConstructorUsedError;
-  bool get displayLabel => throw _privateConstructorUsedError;
-  bool get uninstall =>
-      throw _privateConstructorUsedError; // Android-only options
-  String? get packageName => throw _privateConstructorUsedError;
-  String? get androidFlavor =>
-      throw _privateConstructorUsedError; // iOS-only options
-  String? get bundleId => throw _privateConstructorUsedError;
-  String? get iosFlavor => throw _privateConstructorUsedError;
-  String get scheme => throw _privateConstructorUsedError;
-  String get xcconfigFile => throw _privateConstructorUsedError;
-  String get configuration => throw _privateConstructorUsedError;
-
-  @JsonKey(ignore: true)
-  $TestCommandConfigCopyWith<TestCommandConfig> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $TestCommandConfigCopyWith<$Res> {
-  factory $TestCommandConfigCopyWith(
-          TestCommandConfig value, $Res Function(TestCommandConfig) then) =
-      _$TestCommandConfigCopyWithImpl<$Res, TestCommandConfig>;
-  @useResult
-  $Res call(
-      {List<Device> devices,
-      List<String> targets,
-      Map<String, String> dartDefines,
-      int repeat,
-      bool displayLabel,
-      bool uninstall,
-      String? packageName,
-      String? androidFlavor,
-      String? bundleId,
-      String? iosFlavor,
-      String scheme,
-      String xcconfigFile,
-      String configuration});
-}
-
-/// @nodoc
-class _$TestCommandConfigCopyWithImpl<$Res, $Val extends TestCommandConfig>
-    implements $TestCommandConfigCopyWith<$Res> {
-  _$TestCommandConfigCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? devices = null,
-    Object? targets = null,
-    Object? dartDefines = null,
-    Object? repeat = null,
-    Object? displayLabel = null,
-    Object? uninstall = null,
-    Object? packageName = freezed,
-    Object? androidFlavor = freezed,
-    Object? bundleId = freezed,
-    Object? iosFlavor = freezed,
-    Object? scheme = null,
-    Object? xcconfigFile = null,
-    Object? configuration = null,
-  }) {
-    return _then(_value.copyWith(
-      devices: null == devices
-          ? _value.devices
-          : devices // ignore: cast_nullable_to_non_nullable
-              as List<Device>,
-      targets: null == targets
-          ? _value.targets
-          : targets // ignore: cast_nullable_to_non_nullable
-              as List<String>,
-      dartDefines: null == dartDefines
-          ? _value.dartDefines
-          : dartDefines // ignore: cast_nullable_to_non_nullable
-              as Map<String, String>,
-      repeat: null == repeat
-          ? _value.repeat
-          : repeat // ignore: cast_nullable_to_non_nullable
-              as int,
-      displayLabel: null == displayLabel
-          ? _value.displayLabel
-          : displayLabel // ignore: cast_nullable_to_non_nullable
-              as bool,
-      uninstall: null == uninstall
-          ? _value.uninstall
-          : uninstall // ignore: cast_nullable_to_non_nullable
-              as bool,
-      packageName: freezed == packageName
-          ? _value.packageName
-          : packageName // ignore: cast_nullable_to_non_nullable
-              as String?,
-      androidFlavor: freezed == androidFlavor
-          ? _value.androidFlavor
-          : androidFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      bundleId: freezed == bundleId
-          ? _value.bundleId
-          : bundleId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      iosFlavor: freezed == iosFlavor
-          ? _value.iosFlavor
-          : iosFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      scheme: null == scheme
-          ? _value.scheme
-          : scheme // ignore: cast_nullable_to_non_nullable
-              as String,
-      xcconfigFile: null == xcconfigFile
-          ? _value.xcconfigFile
-          : xcconfigFile // ignore: cast_nullable_to_non_nullable
-              as String,
-      configuration: null == configuration
-          ? _value.configuration
-          : configuration // ignore: cast_nullable_to_non_nullable
-              as String,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$_TestCommandConfigCopyWith<$Res>
-    implements $TestCommandConfigCopyWith<$Res> {
-  factory _$$_TestCommandConfigCopyWith(_$_TestCommandConfig value,
-          $Res Function(_$_TestCommandConfig) then) =
-      __$$_TestCommandConfigCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call(
-      {List<Device> devices,
-      List<String> targets,
-      Map<String, String> dartDefines,
-      int repeat,
-      bool displayLabel,
-      bool uninstall,
-      String? packageName,
-      String? androidFlavor,
-      String? bundleId,
-      String? iosFlavor,
-      String scheme,
-      String xcconfigFile,
-      String configuration});
-}
-
-/// @nodoc
-class __$$_TestCommandConfigCopyWithImpl<$Res>
-    extends _$TestCommandConfigCopyWithImpl<$Res, _$_TestCommandConfig>
-    implements _$$_TestCommandConfigCopyWith<$Res> {
-  __$$_TestCommandConfigCopyWithImpl(
-      _$_TestCommandConfig _value, $Res Function(_$_TestCommandConfig) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? devices = null,
-    Object? targets = null,
-    Object? dartDefines = null,
-    Object? repeat = null,
-    Object? displayLabel = null,
-    Object? uninstall = null,
-    Object? packageName = freezed,
-    Object? androidFlavor = freezed,
-    Object? bundleId = freezed,
-    Object? iosFlavor = freezed,
-    Object? scheme = null,
-    Object? xcconfigFile = null,
-    Object? configuration = null,
-  }) {
-    return _then(_$_TestCommandConfig(
-      devices: null == devices
-          ? _value._devices
-          : devices // ignore: cast_nullable_to_non_nullable
-              as List<Device>,
-      targets: null == targets
-          ? _value._targets
-          : targets // ignore: cast_nullable_to_non_nullable
-              as List<String>,
-      dartDefines: null == dartDefines
-          ? _value._dartDefines
-          : dartDefines // ignore: cast_nullable_to_non_nullable
-              as Map<String, String>,
-      repeat: null == repeat
-          ? _value.repeat
-          : repeat // ignore: cast_nullable_to_non_nullable
-              as int,
-      displayLabel: null == displayLabel
-          ? _value.displayLabel
-          : displayLabel // ignore: cast_nullable_to_non_nullable
-              as bool,
-      uninstall: null == uninstall
-          ? _value.uninstall
-          : uninstall // ignore: cast_nullable_to_non_nullable
-              as bool,
-      packageName: freezed == packageName
-          ? _value.packageName
-          : packageName // ignore: cast_nullable_to_non_nullable
-              as String?,
-      androidFlavor: freezed == androidFlavor
-          ? _value.androidFlavor
-          : androidFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      bundleId: freezed == bundleId
-          ? _value.bundleId
-          : bundleId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      iosFlavor: freezed == iosFlavor
-          ? _value.iosFlavor
-          : iosFlavor // ignore: cast_nullable_to_non_nullable
-              as String?,
-      scheme: null == scheme
-          ? _value.scheme
-          : scheme // ignore: cast_nullable_to_non_nullable
-              as String,
-      xcconfigFile: null == xcconfigFile
-          ? _value.xcconfigFile
-          : xcconfigFile // ignore: cast_nullable_to_non_nullable
-              as String,
-      configuration: null == configuration
-          ? _value.configuration
-          : configuration // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$_TestCommandConfig implements _TestCommandConfig {
-  const _$_TestCommandConfig(
-      {required final List<Device> devices,
-      required final List<String> targets,
-      required final Map<String, String> dartDefines,
-      required this.repeat,
-      required this.displayLabel,
-      required this.uninstall,
-      required this.packageName,
-      required this.androidFlavor,
-      required this.bundleId,
-      required this.iosFlavor,
-      required this.scheme,
-      required this.xcconfigFile,
-      required this.configuration})
-      : _devices = devices,
-        _targets = targets,
-        _dartDefines = dartDefines;
-
-  final List<Device> _devices;
-  @override
-  List<Device> get devices {
-    if (_devices is EqualUnmodifiableListView) return _devices;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_devices);
-  }
-
-  final List<String> _targets;
-  @override
-  List<String> get targets {
-    if (_targets is EqualUnmodifiableListView) return _targets;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_targets);
-  }
-
-  final Map<String, String> _dartDefines;
-  @override
-  Map<String, String> get dartDefines {
-    if (_dartDefines is EqualUnmodifiableMapView) return _dartDefines;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableMapView(_dartDefines);
-  }
-
-  @override
-  final int repeat;
-  @override
-  final bool displayLabel;
-  @override
-  final bool uninstall;
-// Android-only options
-  @override
-  final String? packageName;
-  @override
-  final String? androidFlavor;
-// iOS-only options
-  @override
-  final String? bundleId;
-  @override
-  final String? iosFlavor;
-  @override
-  final String scheme;
-  @override
-  final String xcconfigFile;
-  @override
-  final String configuration;
-
-  @override
-  String toString() {
-    return 'TestCommandConfig(devices: $devices, targets: $targets, dartDefines: $dartDefines, repeat: $repeat, displayLabel: $displayLabel, uninstall: $uninstall, packageName: $packageName, androidFlavor: $androidFlavor, bundleId: $bundleId, iosFlavor: $iosFlavor, scheme: $scheme, xcconfigFile: $xcconfigFile, configuration: $configuration)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$_TestCommandConfig &&
-            const DeepCollectionEquality().equals(other._devices, _devices) &&
-            const DeepCollectionEquality().equals(other._targets, _targets) &&
-            const DeepCollectionEquality()
-                .equals(other._dartDefines, _dartDefines) &&
-            (identical(other.repeat, repeat) || other.repeat == repeat) &&
-            (identical(other.displayLabel, displayLabel) ||
-                other.displayLabel == displayLabel) &&
-            (identical(other.uninstall, uninstall) ||
-                other.uninstall == uninstall) &&
-            (identical(other.packageName, packageName) ||
-                other.packageName == packageName) &&
-            (identical(other.androidFlavor, androidFlavor) ||
-                other.androidFlavor == androidFlavor) &&
-            (identical(other.bundleId, bundleId) ||
-                other.bundleId == bundleId) &&
-            (identical(other.iosFlavor, iosFlavor) ||
-                other.iosFlavor == iosFlavor) &&
-            (identical(other.scheme, scheme) || other.scheme == scheme) &&
-            (identical(other.xcconfigFile, xcconfigFile) ||
-                other.xcconfigFile == xcconfigFile) &&
-            (identical(other.configuration, configuration) ||
-                other.configuration == configuration));
-  }
-
-  @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      const DeepCollectionEquality().hash(_devices),
-      const DeepCollectionEquality().hash(_targets),
-      const DeepCollectionEquality().hash(_dartDefines),
-      repeat,
-      displayLabel,
-      uninstall,
-      packageName,
-      androidFlavor,
-      bundleId,
-      iosFlavor,
-      scheme,
-      xcconfigFile,
-      configuration);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$_TestCommandConfigCopyWith<_$_TestCommandConfig> get copyWith =>
-      __$$_TestCommandConfigCopyWithImpl<_$_TestCommandConfig>(
-          this, _$identity);
-}
-
-abstract class _TestCommandConfig implements TestCommandConfig {
-  const factory _TestCommandConfig(
-      {required final List<Device> devices,
-      required final List<String> targets,
-      required final Map<String, String> dartDefines,
-      required final int repeat,
-      required final bool displayLabel,
-      required final bool uninstall,
-      required final String? packageName,
-      required final String? androidFlavor,
-      required final String? bundleId,
-      required final String? iosFlavor,
-      required final String scheme,
-      required final String xcconfigFile,
-      required final String configuration}) = _$_TestCommandConfig;
-
-  @override
-  List<Device> get devices;
-  @override
-  List<String> get targets;
-  @override
-  Map<String, String> get dartDefines;
-  @override
-  int get repeat;
-  @override
-  bool get displayLabel;
-  @override
-  bool get uninstall;
-  @override // Android-only options
-  String? get packageName;
-  @override
-  String? get androidFlavor;
-  @override // iOS-only options
-  String? get bundleId;
-  @override
-  String? get iosFlavor;
-  @override
-  String get scheme;
-  @override
-  String get xcconfigFile;
-  @override
-  String get configuration;
-  @override
-  @JsonKey(ignore: true)
-  _$$_TestCommandConfigCopyWith<_$_TestCommandConfig> get copyWith =>
-      throw _privateConstructorUsedError;
-}
