+++ packages/patrol_cli/lib/src/features/devices/device_finder.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:dispose_scope/dispose_scope.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:meta/meta.dart';
 import 'package:patrol_cli/src/base/exceptions.dart';
 import 'package:patrol_cli/src/base/logger.dart';
 import 'package:patrol_cli/src/base/process.dart';
