+++ packages/patrol_cli/lib/src/runner/patrol_command_runner.dart
@@ -11,12 +11,12 @@ import 'package:patrol_cli/src/base/constants.dart';
 import 'package:patrol_cli/src/base/exceptions.dart';
 import 'package:patrol_cli/src/base/logger.dart';
 import 'package:patrol_cli/src/base/process.dart';
-import 'package:patrol_cli/src/commands/build_command.dart';
-import 'package:patrol_cli/src/commands/develop_command.dart';
-import 'package:patrol_cli/src/commands/devices_command.dart';
-import 'package:patrol_cli/src/commands/doctor_command.dart';
-import 'package:patrol_cli/src/commands/test_command.dart';
-import 'package:patrol_cli/src/commands/update_command.dart';
+import 'package:patrol_cli/src/commands/build.dart';
+import 'package:patrol_cli/src/commands/develop.dart';
+import 'package:patrol_cli/src/commands/devices.dart';
+import 'package:patrol_cli/src/commands/doctor.dart';
+import 'package:patrol_cli/src/commands/test.dart';
+import 'package:patrol_cli/src/commands/update.dart';
 import 'package:patrol_cli/src/features/devices/device_finder.dart';
 import 'package:patrol_cli/src/features/run_commons/dart_defines_reader.dart';
 import 'package:patrol_cli/src/features/run_commons/test_finder.dart';
