+++ packages/patrol_cli/test/patrol_command_runner_test.dart
@@ -2,8 +2,8 @@ import 'package:args/command_runner.dart';
 import 'package:file/memory.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:patrol_cli/src/base/constants.dart';
+import 'package:patrol_cli/src/base/extensions/command_runner.dart';
 import 'package:patrol_cli/src/base/logger.dart';
-import 'package:patrol_cli/src/common/extensions/command_runner.dart';
 import 'package:patrol_cli/src/runner/patrol_command_runner.dart';
 import 'package:pub_updater/pub_updater.dart';
 import 'package:test/test.dart';
